requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
selenium==4.20.0
seleniumbase==4.26.4
djangorestframework==3.15.1
django-filter==24.2